Now noteBuffer is playing the same loop as before

+++ Rock/Code/Audio/Player.h
@@ -13,6 +13,6 @@
 
 - (void) start;
 
-- (void) setTempo:(NSInteger)tempo;
+- (void) setTempo:(int)tempo;
 
 @end
